requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
opencv-python
numpy
matplotlib
pillow
requests
gdown
rasterio
albumentations
